+++ src/skills.c
@@ -570,8 +570,8 @@ int exp_per_level(CHAR_DATA *ch, int points)
     inc = 500;
 
     if (points < 40)
-	return 1000 * (race_table[ch->race].class_mult[ch->class] ?
-		       race_table[ch->race].class_mult[ch->class]/100 : 1);
+	return 1000 * (pc_race_table[ch->race].class_mult[ch->class] ?
+		       pc_race_table[ch->race].class_mult[ch->class]/100 : 1);
 
     /* processing */
     points -= 40;
@@ -590,7 +590,7 @@ int exp_per_level(CHAR_DATA *ch, int points)
 
     expl += points * inc / 10;  
 
-    return expl * race_table[ch->race].class_mult[ch->class]/100;
+    return expl * pc_race_table[ch->race].class_mult[ch->class]/100;
 }
 
 /* shows all groups, or the sub-members of a group */